ADDITIONAL SUBCLASS - GAMEPLAY GUI
This commit adds a subclass to prevent logical weirdness by having main menu GUI and the actual gameplay GUI at the same level. So now the game play gui can hold the text log, whereas the main menu
can use other things that aren't text logs or to do with gameplay, reducing its bloat. All these files were refactored to move methods, etc, and it works.
